requirements.txt
anthropic>=0.34.0
twilio>=9.0.0
feedparser>=6.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
yfinance>=0.2.40
google-api-python-client>=2.140.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
